consolidating and moving secagent config package
