+++ cmd/security-agent/subcommands/config/config.go
@@ -3,18 +3,30 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build kubeapiserver
+// +build kubeapiserver
+
 package config
 
 import (
 	"fmt"
-
-	"github.com/DataDog/datadog-agent/pkg/config/settings"
-
 	"github.com/spf13/cobra"
+
+	"github.com/DataDog/datadog-agent/cmd/security-agent/command"
+	"github.com/DataDog/datadog-agent/pkg/api/util"
+	"github.com/DataDog/datadog-agent/pkg/config"
+	"github.com/DataDog/datadog-agent/pkg/config/settings"
+	settingshttp "github.com/DataDog/datadog-agent/pkg/config/settings/http"
 )
+
+func Commands(globalParams *command.GlobalParams) []*cobra.Command {
+	cmd := Config(getSettingsClient)
+	return []*cobra.Command{cmd}
+}
 
 // Config returns the main cobra config command.
 func Config(getClient settings.ClientBuilder) *cobra.Command {
+	// TODO: Convert to fx
 	cmd := &cobra.Command{
 		Use:   "config",
 		Short: "Print the runtime configuration of a running agent",
@@ -27,6 +39,44 @@ func Config(getClient settings.ClientBuilder) *cobra.Command {
 	cmd.AddCommand(get(getClient))
 
 	return cmd
+}
+
+func getSettingsClient(cmd *cobra.Command, _ []string) (settings.Client, error) {
+	err := setupConfig(cmd)
+	if err != nil {
+		return nil, err
+	}
+
+	c := util.GetClient(false)
+	apiConfigURL := fmt.Sprintf("https://localhost:%v/agent/config", config.Datadog.GetInt("security_agent.cmd_port"))
+
+	return settingshttp.NewClient(c, apiConfigURL, "security-agent"), nil
+}
+
+func setupConfig(cmd *cobra.Command) error {
+	err := config.SetupLogger(command.LoggerName, config.GetEnvDefault("DD_LOG_LEVEL", "off"), "", "", false, true, false)
+	if err != nil {
+		fmt.Printf("Cannot setup logger, exiting: %v\n", err)
+		return err
+	}
+
+	return util.SetAuthToken()
+}
+
+func showRuntimeConfiguration(getClient settings.ClientBuilder, cmd *cobra.Command, args []string) error {
+	c, err := getClient(cmd, args)
+	if err != nil {
+		return err
+	}
+
+	runtimeConfig, err := c.FullConfig()
+	if err != nil {
+		return err
+	}
+
+	fmt.Println(runtimeConfig)
+
+	return nil
 }
 
 // listRuntime returns a cobra command to list the settings that can be changed at runtime.
@@ -59,43 +109,6 @@ func get(getClient settings.ClientBuilder) *cobra.Command {
 		Long:  ``,
 		RunE:  func(cmd *cobra.Command, args []string) error { return getConfigValue(getClient, cmd, args) },
 	}
-}
-
-func showRuntimeConfiguration(getClient settings.ClientBuilder, cmd *cobra.Command, args []string) error {
-	c, err := getClient(cmd, args)
-	if err != nil {
-		return err
-	}
-
-	runtimeConfig, err := c.FullConfig()
-	if err != nil {
-		return err
-	}
-
-	fmt.Println(runtimeConfig)
-
-	return nil
-}
-
-func listRuntimeConfigurableValue(getClient settings.ClientBuilder, cmd *cobra.Command, args []string) error {
-	c, err := getClient(cmd, args)
-	if err != nil {
-		return err
-	}
-
-	settingsList, err := c.List()
-	if err != nil {
-		return err
-	}
-
-	fmt.Println("=== Settings that can be changed at runtime ===")
-	for setting, details := range settingsList {
-		if !details.Hidden {
-			fmt.Printf("%-30s %s\n", setting, details.Description)
-		}
-	}
-
-	return nil
 }
 
 func setConfigValue(getClient settings.ClientBuilder, cmd *cobra.Command, args []string) error {
@@ -138,6 +151,27 @@ func getConfigValue(getClient settings.ClientBuilder, cmd *cobra.Command, args [
 	}
 
 	fmt.Printf("%s is set to: %v\n", args[0], value)
+
+	return nil
+}
+
+func listRuntimeConfigurableValue(getClient settings.ClientBuilder, cmd *cobra.Command, args []string) error {
+	c, err := getClient(cmd, args)
+	if err != nil {
+		return err
+	}
+
+	settingsList, err := c.List()
+	if err != nil {
+		return err
+	}
+
+	fmt.Println("=== Settings that can be changed at runtime ===")
+	for setting, details := range settingsList {
+		if !details.Hidden {
+			fmt.Printf("%-30s %s\n", setting, details.Description)
+		}
+	}
 
 	return nil
 }
